+++ tools/ciprobe/main.go
@@ -17,6 +17,8 @@ import (
 	"regexp"
 	"strings"
 	"time"
+
+	"scriptureforge/tools/probehttp"
 )
 
 type config struct {
@@ -89,7 +91,7 @@ func parseFlags() config {
 }
 
 func run(cfg config, output io.Writer) error {
-	return runWithClient(cfg, output, &http.Client{Timeout: cfg.Timeout})
+	return runWithClient(cfg, output, probehttp.NewClient(cfg.Timeout))
 }
 
 func runWithClient(cfg config, output io.Writer, client *http.Client) error {
@@ -116,7 +118,7 @@ func runWithClient(cfg config, output io.Writer, client *http.Client) error {
 		}
 	}
 	if client == nil {
-		client = &http.Client{Timeout: cfg.Timeout}
+		client = probehttp.NewClient(cfg.Timeout)
 	}
 
 	probes := []probeResult{